Extract pathway constants to shared module
